expr: shrink ColumnOrder from 16 to 8 bytes via usize→u32 for column field
Change `ColumnOrder::column` from `usize` to `u32`, halving the struct
size from 16 to 8 bytes. No table can have more than ~4 billion columns,
so u32 is more than sufficient.

Unlike boxing optimizations, this has zero overhead — no extra indirection
or allocation. Every `ColumnOrder` in memory is simply half the size.

This benefits all ORDER BY, TopK, and window function operations since
ColumnOrder is stored in Vec/Box<[T]> arrays throughout the planner,
optimizer, and execution layers.

+++ src/adapter/src/active_compute_sink.rs
@@ -246,7 +246,7 @@ impl ActiveSubscribe {
                             });
                             let datums = datum_vec.borrow_with(row);
                             for column_order in order_by_keys {
-                                packer.push(datums[column_order.column]);
+                                packer.push(datums[column_order.column as usize]);
                             }
                             if debezium {
                                 for _ in 0..value_columns {
@@ -254,7 +254,7 @@ impl ActiveSubscribe {
                                 }
                             }
                             for idx in 0..self.arity {
-                                if !order_by_keys.iter().any(|co| co.column == idx) {
+                                if !order_by_keys.iter().any(|co| co.column as usize == idx) {
                                     packer.push(datums[idx]);
                                 }
                             }
@@ -264,11 +264,11 @@ impl ActiveSubscribe {
                             packer.push(Datum::String("delete"));
                             let datums = datum_vec.borrow_with(&start.1);
                             for column_order in order_by_keys {
-                                packer.push(datums[column_order.column]);
+                                packer.push(datums[column_order.column as usize]);
                             }
                             if debezium {
                                 for idx in 0..self.arity {
-                                    if !order_by_keys.iter().any(|co| co.column == idx) {
+                                    if !order_by_keys.iter().any(|co| co.column as usize == idx) {
                                         packer.push(datums[idx]);
                                     }
                                 }
@@ -284,17 +284,17 @@ impl ActiveSubscribe {
                             let old_datums = old_datum_vec.borrow_with(old_row);
 
                             for column_order in order_by_keys {
-                                packer.push(datums[column_order.column]);
+                                packer.push(datums[column_order.column as usize]);
                             }
                             if debezium {
                                 for idx in 0..self.arity {
-                                    if !order_by_keys.iter().any(|co| co.column == idx) {
+                                    if !order_by_keys.iter().any(|co| co.column as usize == idx) {
                                         packer.push(old_datums[idx]);
                                     }
                                 }
                             }
                             for idx in 0..self.arity {
-                                if !order_by_keys.iter().any(|co| co.column == idx) {
+                                if !order_by_keys.iter().any(|co| co.column as usize == idx) {
                                     packer.push(datums[idx]);
                                 }
                             }
@@ -304,7 +304,7 @@ impl ActiveSubscribe {
                             packer.push(Datum::String("key_violation"));
                             let datums = datum_vec.borrow_with(&start.1);
                             for column_order in order_by_keys {
-                                packer.push(datums[column_order.column]);
+                                packer.push(datums[column_order.column as usize]);
                             }
                             if debezium {
                                 for _ in 0..(self.arity - order_by_keys.len()) {
